[refactor] 패키지 구조 변경
- DispatcherServlet, Adapters, Mappings 등 이동
